move common folder into fsbp-fix
