+++ fsbp-fix/common/aws.go
