+++ fsbp-fix/common/math.go
