+++ fsbp-fix/go.mod
@@ -2,14 +2,13 @@ module github.com/guardian/fsbp-tools/fsbp-fix
 
 go 1.22.1
 
-replace github.com/guardian/fsbp-tools/common => ../common
-
 require (
 	github.com/aws/aws-sdk-go-v2/service/cloudformation v1.51.3
 	github.com/aws/aws-sdk-go-v2/service/s3 v1.55.1
 	github.com/aws/aws-sdk-go-v2/service/securityhub v1.49.2
-	github.com/guardian/fsbp-tools/common v0.0.0-00010101000000-000000000000
 	github.com/aws/aws-sdk-go-v2/service/ec2 v1.175.0
+	github.com/aws/aws-sdk-go-v2/config v1.27.18
+	github.com/aws/aws-sdk-go-v2/service/sts v1.28.12
 )
 
 require (
